requirements.txt
praw
tweepy
python-dotenv
azure-keyvault-secrets
azure-identity
azure-functions
azure-storage-blob